api and class Note added2
