+++ .gitignore
@@ -3,3 +3,4 @@ __pycache__/
 env/
 config.py
 *.db
+.idea/
